Move deprecated viewports to deprecated folder
